Flatten accounts index operations
Rather than operating on the entire database, the functions in the
accounts index now only operate on the accounts list itself. This means
the accounts index will need to have less knowledge about how it is
stored going forward.

The .all() function is the only remaining exception because the
transaction schema depends on it, but it should be removed as soon as
that is no longer the case.