Flatten transactions index operations
The transactions index now only operates on the transactions list
itself. To facilitate this change, the schema had to be updated to no
longer take in the global database to verify the existence of accounts
and payees. Instead, casting a transaction will only verify that the
account and payee ids are valid types, and existence checks will be left
to the caster.