[AddAccount] Rudimentary starting balance support
This doesn't account for many edge cases and input possibilities, but it
does require a starting balance in order to create a new account.

The starting balance is itself a normal transaction, though all of the
details other than amount are defaulted by the system rather than
inputted by the user.