[AddAccount] Datepicker for initial balance date
This removes much of the existing date parsing logic in favor of using
the `moment` module, though it would still be ideal if date handling
logic were abstracted behind a single internal module in a future
commit.

Webpack css/style loaders are also added to the project for the first
time.