[Styling] Base webpack config for styles
This uses two libraries for styles, Sass and postcss/autoprefixer. Styles
can be written in Sass and transpiled into CSS. Rather than using
'mixins' to handle vendor prefixes in Sass, we're running autoprefixer
on the transpiled css to automatically inject vendor prefixes where
they're needed. This also includes base styling for the application,
which can be found in 'base.css' and 'core.css'.